pricelab/requirements.txt
asgiref==3.7.2
asn1crypto==1.5.1
bootstrap4==0.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
Django==5.0.1
django-admin-interface==0.28.3
django-background-tasks==1.2.5
django-colorfield==0.11.0
django-compat==1.0.15
django-crispy-forms==2.1
filelock==3.13.1
fonttools==4.47.2
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
patsy==0.5.6
pillow==10.2.0
platformdirs==3.11.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.51
sqlparse==0.4.4
statsmodels==0.14.1
text-unidecode==1.3
threadpoolctl==3.2.0
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
